refactor: study에서 round에 대한 의존성 제거
