+++ backend/src/main/java/com/yigongil/backend/domain/study/StudyFinishedEvent.java
@@ -1,5 +1,5 @@
 package com.yigongil.backend.domain.study;
 
-public record StudyFinishedEvent(Long studyId, Integer roundExperience) {
+public record StudyFinishedEvent(Long studyId, Integer roundExperience, Integer minimumWeeks) {
 
 }
