+++ backend/src/main/java/com/yigongil/backend/domain/study/studymember/Role.java
@@ -1,4 +1,4 @@
-package com.yigongil.backend.domain.studymember;
+package com.yigongil.backend.domain.study.studymember;
 
 import java.util.Arrays;
 import lombok.Getter;
