+++ backend/src/main/java/com/yigongil/backend/domain/study/studymember/RoleConverter.java
@@ -1,4 +1,4 @@
-package com.yigongil.backend.domain.studymember;
+package com.yigongil.backend.domain.study.studymember;
 
 import org.springframework.core.convert.converter.Converter;
