+++ backend/src/main/java/com/yigongil/backend/domain/study/studymember/StudyResult.java
@@ -1,4 +1,4 @@
-package com.yigongil.backend.domain.studymember;
+package com.yigongil.backend.domain.study.studymember;
 
 public enum StudyResult {
