+++ backend/src/main/java/com/yigongil/backend/domain/study/studyquery/StudyQueryRepository.java
@@ -2,7 +2,7 @@ package com.yigongil.backend.domain.study.studyquery;
 
 import com.yigongil.backend.domain.study.ProcessingStatus;
 import com.yigongil.backend.domain.study.Study;
-import com.yigongil.backend.domain.studymember.Role;
+import com.yigongil.backend.domain.study.studymember.Role;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
