+++ backend/src/main/java/com/yigongil/backend/query/profile/ProfileService.java
@@ -5,8 +5,8 @@ import com.yigongil.backend.domain.member.domain.MemberRepository;
 import com.yigongil.backend.domain.round.RoundRepository;
 import com.yigongil.backend.domain.study.Study;
 import com.yigongil.backend.domain.study.StudyService;
-import com.yigongil.backend.domain.studymember.StudyMember;
-import com.yigongil.backend.domain.studymember.StudyMemberRepository;
+import com.yigongil.backend.domain.study.studymember.StudyMember;
+import com.yigongil.backend.domain.study.studymember.StudyMemberRepository;
 import com.yigongil.backend.response.FinishedStudyResponse;
 import com.yigongil.backend.response.ProfileResponse;
 import java.util.List;
