+++ backend/src/test/java/com/yigongil/backend/fixture/StudyFixture.java
@@ -1,19 +1,14 @@
 package com.yigongil.backend.fixture;
 
 import com.yigongil.backend.domain.member.domain.Member;
-import com.yigongil.backend.domain.round.Round;
 import com.yigongil.backend.domain.study.ProcessingStatus;
 import com.yigongil.backend.domain.study.Study;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 public enum StudyFixture {
 
     자바_스터디_모집중(null, LocalDateTime.now(), "자바", "스터디소개", ProcessingStatus.RECRUITING, 4, 2, 4),
-    자바_스터디_모집중_정원_2(null, LocalDateTime.now(), "자바", "스터디소개", ProcessingStatus.RECRUITING, 2, 2, 4)
-    ;
+    자바_스터디_모집중_정원_2(null, LocalDateTime.now(), "자바", "스터디소개", ProcessingStatus.RECRUITING, 2, 2, 4);
 
     private final Long id;
     private final String name;
@@ -57,21 +52,6 @@ public enum StudyFixture {
                     .numberOfMaximumMembers(numberOfMaximumMember)
                     .meetingDaysCountPerWeek(meetingDaysPerWeek)
                     .minimumWeeks(minimumWeeks)
-                    .build();
-    }
-
-    public Study toStudyWithRounds(RoundFixture... roundFixtures) {
-        List<Round> rounds = new ArrayList<>(Arrays.stream(roundFixtures)
-                                                   .map(RoundFixture::toRound)
-                                                   .toList());
-
-        return Study.builder()
-                    .id(id)
-                    .name(name)
-                    .introduction(introduction)
-                    .processingStatus(processingStatus)
-                    .numberOfMaximumMembers(numberOfMaximumMember)
-                    .rounds(rounds)
                     .build();
     }
 }
